fix: LoginContext 구조 개선
